refactor: rename the adapter mocks package
